component pattern with angularjs
added an "ngListItem" component.

explanation of bindings:

- index (<) - used for tracking which todo to modify
- todo (<) - one way binding for todo object
- onCheck (&) - parent specifies function params
- onEdit (&) - parent specifies function params
- onRemove (&) - parent specifies function params

each component is in control of its own "editing" todo which
is why the input todo gets copied whenever the component
is in edit mode.

i'm on the fence on whether it's better to use "&" or "<"
bindings for callbacks.

ALSO - i updated the mainCtrl code to update the todos
array differently and THEN set $scope.todos = todoStorage.todos
so that things get applied as expected without modification
and having to avoid using any $scope.apply()s