use react components
JSX is useful, so i added the babel cli tool to watch for changes
and compile it into the "dist/" folder.

there's a bit of a trick to get the react components to correctly
call it's callbacks so that the parent knows to "rerender" or apply
the changes. the solution is to wrap each of the callback functions
in a "$scope.apply()". we can use some naming scheme to see if a prop
is a function and starts with "on" that we can wrap the scope call.